% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derfinderHelper-package.R
\docType{package}
\name{derfinderHelper-package}
\alias{derfinderHelper-package}
\title{derfinder helper package}
\description{
Helper package for speeding up the derfinder package when using multiple
cores.
}
\author{
Leonardo Collado-Torres \href{mailto:lcolladotor@gmail.com}{lcolladotor@gmail.com}
}
\keyword{package}
