% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{derfinderHelper-package}
\alias{derfinderHelper-package}
\title{derfinder helper package}
\description{
Helper package for speeding up the derfinder package when using multiple
cores.
}
\author{
Leonardo Collado-Torres <lcollado@jhu.edu>
}
\keyword{package}

